requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
outcome==1.1.0
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
wsproto==1.0.0
zipp==3.7.0
